telegram-bot/requirements.txt
pytelegrambotapi
python-decouple
requests
